requirements.txt
altair==5.5.0
backtesting
google-generativeai
matplotlib==3.10.8
mplfinance==0.12.10b0
numpy
pandas
pandas_ta
Pillow==12.0.0
protobuf
psutil==7.1.3
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
streamlit==1.51.0
websockets==10.4
yfinance==0.2.40
colorama